Add ability to open login window first, and then not reopen it after logged in
