+++ app/src/androidTest/java/ch/epfl/favo/MainActivityTest.java
@@ -22,6 +22,14 @@ public class MainActivityTest {
     @Rule
     public final ActivityTestRule<MainActivity> mActivityRule =
             new ActivityTestRule<>(MainActivity.class);
+    @Test
+    public void sampleTest(){
+        onView(withId(R.id.mainGoButton)).perform(click());
+    }
+    //@Test
+    //public void sampleTest(){
+    //    onView(withId(R.id.mainGoButton)).perform(click());
+    //}
     /**
      * Sample Test function underneath
      */
