--- app/src/main/java/ch/epfl/favo/presenter/AccountInfoActivity.java
@@ -1,8 +1,0 @@
-package ch.epfl.favo.presenter;
-
-/**
- * Contains logic for account info view (3rd tab)
- * TODO: Implement account info (Andrea)
- */
-public class AccountInfoActivity {
-}
