--- app/src/main/java/ch/epfl/favo/presenter/DisplayMessageActivity.java
@@ -1,30 +1,0 @@
-package ch.epfl.favo.presenter;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.widget.TextView;
-
-import ch.epfl.favo.R;
-import ch.epfl.favo.presenter.MainActivity;
-
-/**
- * Example presenter class from Bootcamp
- */
-public class DisplayMessageActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_display_message);
-
-        // Get the Intent that started this activity and extract the string
-        Intent intent = getIntent();
-        String message = intent.getStringExtra(MainActivity.EXTRA_MESSAGE);
-
-        // Capture the layout's TextView and set the string as its text
-        TextView textView = findViewById(R.id.greetingMessage);
-        textView.setText(getString(R.string.greetingMes, message));
-    }
-}
