--- app/src/main/java/ch/epfl/favo/presenter/FavorListActivity.java
@@ -1,8 +1,0 @@
-package ch.epfl.favo.presenter;
-
-/**
- * Will contain the list view from the second tab
- * TODO: Implement favor list activity
- */
-public class FavorListActivity {
-}
