--- app/src/main/java/ch/epfl/favo/presenter/MapActivity.java
@@ -1,7 +1,0 @@
-package ch.epfl.favo.presenter;
-
-/**
- * TODO: Implement map logic HUajian
- */
-public class MapActivity {
-}
