--- app/src/main/java/ch/epfl/favo/presenter/PostFavorActivity.java
@@ -1,9 +1,0 @@
-package ch.epfl.favo.presenter;
-
-/**
- * This class should allow us to fetch all the information required to specify a favor.
- * TODO: David implement favor request interface
- */
-public class PostFavorActivity {
-
-}
