--- app/src/main/java/ch/epfl/favo/presenter/SignupActivity.java
@@ -1,7 +1,0 @@
-package ch.epfl.favo.presenter;
-
-/**
- * TODO: Implement signup logic
- */
-public class SignupActivity {
-}
