+++ app/src/main/res/layout/activity_login.xml
@@ -5,9 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".presenter.MainActivity">
-
-
+    tools:context=".presenter.LoginActivity">
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -59,7 +57,9 @@
             android:layout_marginBottom="10dp"
             android:padding="20dp"
             android:text="Log in"
+            android:onClick="login"
             android:textAllCaps="true"
             android:textAlignment="center"/>
     </LinearLayout>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
